% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAssignments}
\alias{addAssignments}
\alias{addAssignments,Analysis,Assignment-method}
\title{Add molecular formula assignments}
\usage{
addAssignments(analysis, assignment)

\S4method{addAssignments}{Analysis,Assignment}(analysis, assignment)
}
\arguments{
\item{analysis}{S4 object of class \code{Analysis}}

\item{assignmwnt}{S4 object of class \code{Assignment}}
}
\value{
An S4 object of class \code{Analysis}
}
\description{
Add molecular formaul assignments to an \code{Analysis} class object.
}
\details{
Molecular formula assignments are added to the \code{Analysis} pre-treated data
}
