% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectBatchDiff}
\alias{detectBatchDiff}
\alias{detectBatchDiff,Binalysis-method}
\alias{detectBatchDiff,MetaboProfile-method}
\title{Detect batch/block differences}
\usage{
detectBatchDiff(x, by = "block", pthresh = 0.05)

\S4method{detectBatchDiff}{Binalysis}(x, by = "block", pthresh = 0.05)

\S4method{detectBatchDiff}{MetaboProfile}(x, by = "block", pthresh = 0.05)
}
\arguments{
\item{x}{object of S4 class \code{Binalysis} or \code{MetaboProfile}}

\item{by}{info class column to use for batch/block information}

\item{pthresh}{p-value threshold for significance}
}
\value{
If no differences between batches are found then \code{NULL} is returned. If significant differences are found then a tibble is returned containing the ANOVA results for each ionisation mode and showing whether batch correction is needed.
}
\description{
Detect batch/block differences within analytical runs for each ionisation mode.
}
\details{
Analysis of Variance (ANOVA) is used to detect differences in total ion count (TIC) averages between batches/blocks.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes',ask = FALSE)

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

## Detect batch differences
batch_diff <- detectBatchDiff(analysis)

## Display batch differences
batch_diff
}
