% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detect.R
\docType{methods}
\name{detectMissInjections}
\alias{detectMissInjections}
\alias{detectMissInjections,Binalysis-method}
\alias{detectMissInjections,MetaboProfile-method}
\title{detectMissInjections}
\usage{
detectMissInjections(x, idx = "fileOrder")

\S4method{detectMissInjections}{Binalysis}(x, idx = "fileOrder")

\S4method{detectMissInjections}{MetaboProfile}(x, idx = "fileOrder")
}
\description{
detect miss injected samples
}
