% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectModellingParameters}
\alias{detectModellingParameters}
\alias{detectModellingParameters,Binalysis-method}
\alias{detectModellingParameters,MetaboProfile-method}
\alias{detectModellingParameters,Analysis-method}
\title{Detect modelling parameters}
\usage{
detectModellingParameters(x, ...)

\S4method{detectModellingParameters}{Binalysis}(x)

\S4method{detectModellingParameters}{MetaboProfile}(x)

\S4method{detectModellingParameters}{Analysis}(x, type = "pre-treated", cls = "class")
}
\arguments{
\item{x}{S4 object of class \code{Binalysis},\code{MetaboProfile} or \code{Analysis}}

\item{...}{arguments to pass to the appropriate method}

\item{type}{detect parameters for \code{raw} or \code{pre-treated} data for \code{Analysis} class}

\item{cls}{sample information column to use for modelling}
}
\value{
S4 object of class \code{AnalysisParameters}
}
\description{
Detect modelling parameters for \code{Binalysis}, \code{MetaboProfile} or \code{Analysis} S4 classes.
}
\examples{
library(metaboData)
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

## Detect modelling parameters
modelling_parameters <- detectModellingParameters(analysis)

modelling_parameters
}
