% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detect.R
\docType{methods}
\name{detectPairwises}
\alias{detectPairwises}
\alias{detectPairwises,Analysis-method}
\title{detectPairwises}
\usage{
detectPairwises(x, cls, type)

\S4method{detectPairwises}{Analysis}(x, cls, type)
}
\description{
detectPairwises
}
