% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectPretreatmentParameters}
\alias{detectPretreatmentParameters}
\alias{detectPretreatmentParameters,Binalysis-method}
\alias{detectPretreatmentParameters,MetaboProfile-method}
\title{Detect pre-treatment parameters}
\usage{
detectPretreatmentParameters(
  x,
  cls = "class",
  QCidx = "QC",
  miss_injections = TRUE,
  batch_correction = TRUE,
  threshold = 25
)

\S4method{detectPretreatmentParameters}{Binalysis}(
  x,
  cls = "class",
  QCidx = "QC",
  miss_injections = TRUE,
  batch_correction = TRUE,
  threshold = 25
)

\S4method{detectPretreatmentParameters}{MetaboProfile}(
  x,
  cls = "class",
  QCidx = "QC",
  miss_injections = TRUE,
  batch_correction = TRUE,
  threshold = 25
)
}
\arguments{
\item{x}{S4 object of class \code{Binalysis}, \code{MetaboProfile} or \code{AnalysisData}}

\item{cls}{the name of the  sample information table column containing the sample class information}

\item{QCidx}{QC sample class label}

\item{miss_injections}{TRUE/FALSE. Detect the presence of possible miss injections and include parameters to remove these if necessary.}

\item{batch_correction}{TRUE/FALSE. Detect if a batch correction is necessary and include parameters to perform this if necessary.}

\item{threshold}{the percentage of the median TIC below which samples will be considered miss injections. This will be ignored if \code{miss_injections = FALSE}.}
}
\value{
An object of S4 class \code{AnalysisParameters}
}
\description{
Detect suitable pre-treatment parameters for \code{Binalysis} or \code{MetaboProfile} class objects.
}
\examples{
## Retreive example file paths and sample information 
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') 

sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes') 

## Detect spectral binning parameters
bp <- binneR::detectParameters(file_paths)

## Perform spectral binning
bd <- binneR::binneRlyse(file_paths,sample_information,bp)

## Detect pre-treatment parameters
pp <- detectPretreatmentParameters(bd) 

pp
}
