% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{export,Binalysis-method}
\alias{export,Analysis-method}
\alias{export,Assignment-method}
\title{Export results}
\usage{
export(analysis, outPath = ".")

\S4method{export}{Binalysis}(analysis, outPath = ".")

\S4method{export}{Analysis}(analysis, outPath = ".")

\S4method{export}{Assignment}(analysis, outPath = ".")
}
\arguments{
\item{analysis}{S4 object of class \code{Binalysis}, \code{MetaboProfile}, \code{Analysis} or \code{Assignment},}

\item{outPath}{directory path to export to.}
}
\description{
Export data tables from \code{Binalysis},\code{MetaboProfile}, \code{Analysis} and \code{Assignment} classes.
}
