% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{featureSummary}
\alias{featureSummary}
\alias{featureSummary,Binalysis-method}
\alias{featureSummary,MetaboProfile-method}
\title{Summarise \code{m/z} features}
\usage{
featureSummary(x)

\S4method{featureSummary}{Binalysis}(x)

\S4method{featureSummary}{MetaboProfile}(x)
}
\arguments{
\item{x}{an object of S4 class \code{Binalysis} or \code{MetaboProfile}}
}
\value{
A tibble containing feature summaries.
}
\description{
Summarise spectrally processed \emph{m/z} features.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

featureSummary(analysis)
}
