% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCorrelations.R
\name{filterCorrelations}
\alias{filterCorrelations}
\title{filterCorrelations}
\usage{
filterCorrelations(correlations, rthresh = 0.7, n = 1e+05,
  rIncrement = 0.01, nIncrement = 20000)
}
\description{
filterCorrelations
}
