% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRSD,Binalysis-method}
\alias{plotRSD,Binalysis-method}
\alias{plotRSD,MetaboProfile-method}
\title{Plot RSD distributions}
\usage{
\S4method{plotRSD}{Binalysis}(analysis, cls = "class")

\S4method{plotRSD}{MetaboProfile}(analysis, cls = "class")
}
\arguments{
\item{analysis}{an object of S4 class \code{Binalysis} or \code{MetaboProfile}}

\item{cls}{the info column to use for class labels}
}
\value{
A list of RSD distribution plots for each ionisation mode..
}
\description{
Plot relative standard deviation (RSD) distributions of spectrally processed data.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

## Plot RSD distributions
plotRSD(analysis)
}
