% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/preTreatModes.R
\docType{methods}
\name{preTreatModes}
\alias{preTreatModes}
\alias{preTreatModes,Binalysis-method}
\title{preTreatModes}
\usage{
preTreatModes(binalysis, parameters)

\S4method{preTreatModes}{Binalysis}(binalysis, parameters)
}
\description{
preTreatModes
}
