% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\alias{reduce,Analysis-method}
\alias{reduce,AnalysisData-method}
\title{Reduce an analysis by removing isotopic, adduct or unknown features}
\usage{
reduce(
  x,
  isotopes = TRUE,
  adducts = TRUE,
  unknowns = FALSE,
  isotopic_adducts = c("[M+Cl37]1-", "[M+K41]1+")
)

\S4method{reduce}{Analysis}(
  x,
  isotopes = TRUE,
  adducts = TRUE,
  unknowns = FALSE,
  isotopic_adducts = c("[M+Cl37]1-", "[M+K41]1+")
)

\S4method{reduce}{AnalysisData}(
  x,
  isotopes = TRUE,
  adducts = TRUE,
  unknowns = FALSE,
  isotopic_adducts = c("[M+Cl37]1-", "[M+K41]1+")
)
}
\arguments{
\item{x}{S4 class of class \code{Analysis} or \code{AnalysisData}}

\item{isotopes}{TRUE/FALSE remove isotopic features.}

\item{adducts}{TRUE/FALSE remove multiple adduct features.}

\item{unknowns}{TRUE/FALSE remove unassigned features.}

\item{isotopic_adducts}{a vector of additional isotopic adducts to remove if argument \code{isotopes = TRUE}}
}
\value{
S4 object of class \code{Analysis} or \code{AnalysisData} with features reduced
}
\description{
Reduce an analysis by removing isotopic, adduct or unknown features.
}
\details{
If argument \code{isotopes = TRUE}, all isotopic features are removed. If argument \code{adducts = TRUE}, the feature with the maximum intensity for each molecular formula is retained.
}
\examples{
## Assign molecular formulas
p <- assignments::assignmentParameters('FIE')

assignment <- assignments::assignMFs(assignments::feature_data,p)

## Retrieve assigned data
assigned_data <- metabolyseR::analysisData(
 assignments::assignedData(assignment),
 tibble::tibble(sample = seq_len(nrow(assignments::feature_data)))
 )

reduced_data <- metaboMisc::reduce(assigned_data)

reduced_data
}
