% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{suitableParallelPlan}
\alias{suitableParallelPlan}
\title{Generate a suitable parallel processing plan}
\usage{
suitableParallelPlan(strategy = NULL, workers = NULL)
}
\arguments{
\item{strategy}{the arallel strategy to use. See \code{?future::plan} for details. If \code{NULL} (the default), an appropriate strategy will be detected.}

\item{workers}{The number of workers to use. If \code{NULL}, the selected number will be three quarterss of the available system workers.}
}
\description{
Generate a suitable parallel processing future plan for your session. See \code{?future::plan} for more details on future plans.
}
\examples{
\dontrun{
suitableParallelPlan()
}
}
