% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neat.R
\name{theme_neat}
\alias{theme_neat}
\title{theme_neat}
\usage{
theme_neat(base_size = 12, base_family = "Ubunbu")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Neat theme for ggplot2
}
