% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAssignments}
\alias{addAssignments}
\alias{addAssignments,Analysis,Assignment-method}
\title{Add molecular formula assignments}
\usage{
addAssignments(analysis, assignment)

\S4method{addAssignments}{Analysis,Assignment}(analysis, assignment)
}
\arguments{
\item{analysis}{an object of S4 class \code{Analysis}}

\item{assignment}{an object of S4 class \code{Assignment} that includes assignemnts for the \emph{m/z} features specified for argument \code{analysis}}
}
\value{
An object of S4 class \code{Analysis} that includes the molecular formulas in the feature names for the pre-treated data.
}
\description{
Add molecular formula assignments to an \code{Analysis} S4 class object.
}
