% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/addAssignments.R
\docType{methods}
\name{addAssignments}
\alias{addAssignments}
\alias{addAssignments,Analysis,Assignment-method}
\alias{addAssignments,AnalysisData,Assignment-method}
\title{addAssignments}
\usage{
addAssignments(analysis, assignment)

\S4method{addAssignments}{Analysis,Assignment}(analysis, assignment)

\S4method{addAssignments}{AnalysisData,Assignment}(analysis, assignment)
}
\arguments{
\item{analysis}{S4 object of class Analysis or AnalysisData}

\item{assignment}{S4 object of class Assignment}
}
\description{
add molecular formula assignments to an Analysis object.
}
