% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSampleInfo}
\alias{convertSampleInfo}
\title{Convert sample information to a workflow compatible format}
\usage{
convertSampleInfo(sample_info)
}
\arguments{
\item{sample_info}{tibble containing sample information as returned from \code{grover::runInfo}}
}
\value{
A tibble containing converted sample information.
}
\description{
Convert sample information from \code{grover::runInfo} to a format compatible with \code{binneR} and \code{profilePro}.
}
