% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSampleInfo}
\alias{convertSampleInfo}
\title{Convert sample information to a workflow compatible format}
\usage{
convertSampleInfo(sample_info, gzip_ext = TRUE)
}
\arguments{
\item{sample_info}{tibble containing sample information as returned from \code{grover::runInfo}}

\item{gzip_ext}{add \code{.gz} extension to file names}
}
\value{
A tibble containing converted sample information.
}
\description{
Convert sample information from \code{grover::runInfo} to a format compatible with \code{binneR} and \code{profilePro}.
}
