% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detect.R
\docType{methods}
\name{detectBatchDiff}
\alias{detectBatchDiff}
\alias{detectBatchDiff,Binalysis-method}
\alias{detectBatchDiff,MetaboProfile-method}
\title{detectBatchDiff}
\usage{
detectBatchDiff(x, by = "block", pthresh = 0.05)

\S4method{detectBatchDiff}{Binalysis}(x, by = "block", pthresh = 0.05)

\S4method{detectBatchDiff}{MetaboProfile}(x, by = "block",
  pthresh = 0.05)
}
\description{
detectBatchDiff
}
