% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detect.R
\name{detectMissInjections}
\alias{detectMissInjections}
\alias{detectMissInjections,Binalysis-method}
\alias{detectMissInjections,MetaboProfile-method}
\title{detectMissInjections}
\usage{
detectMissInjections(x, idx = "fileOrder")

\S4method{detectMissInjections}{Binalysis}(x, idx = "fileOrder")

\S4method{detectMissInjections}{MetaboProfile}(x, idx = "fileOrder")
}
\arguments{
\item{x}{object of class Binalysis or MetaboProfile}

\item{idx}{info column to use for sample indexes}
}
\description{
detect miss injected samples
}
\examples{
\dontrun{
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

miss_injections <- detectMissInjections(analysis)

miss_injections$missInjections
}
}
