% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectMissInjections}
\alias{detectMissInjections}
\alias{detectMissInjections,Binalysis-method}
\alias{detectMissInjections,MetaboProfile-method}
\title{Detect miss injections}
\usage{
detectMissInjections(x, idx = "injOrder", threshold = 25)

\S4method{detectMissInjections}{Binalysis}(x, idx = "injOrder", threshold = 25)

\S4method{detectMissInjections}{MetaboProfile}(x, idx = "injOrder", threshold = 25)
}
\arguments{
\item{x}{object of S4 class \code{Binalysis} or \code{MetaboProfile}}

\item{idx}{the sample information column to use for sample indexes}

\item{threshold}{the percentage of the median TIC below which samples will be considered miss injections.}
}
\value{
A list containing the name of the sample information column used to index the miss injections and a vector of miss injection indexes.
}
\description{
Detect miss injected samples or samples with a total ion count below a percentage threshold of the median.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

## Detect miss injections
miss_injections <- detectMissInjections(analysis)

## Display detected miss injections
miss_injections
}
