% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/detect.R
\name{detectPairwises}
\alias{detectPairwises}
\alias{detectPairwises,Analysis-method}
\title{detectPairwises}
\usage{
detectPairwises(x, cls, type)

\S4method{detectPairwises}{Analysis}(x, cls, type)
}
\arguments{
\item{x}{object of class Analysis}

\item{cls}{info column to use for class information}

\item{type}{type of analysis (classification or featureSelection)}
}
\description{
Detect availble pairwise comparisons
}
