% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectPretreatmentParameters}
\alias{detectPretreatmentParameters}
\alias{detectPretreatmentParameters,Binalysis-method}
\alias{detectPretreatmentParameters,MetaboProfile-method}
\title{Detect pre-treatment parameters}
\usage{
detectPretreatmentParameters(x)

\S4method{detectPretreatmentParameters}{Binalysis}(x)

\S4method{detectPretreatmentParameters}{MetaboProfile}(x)
}
\arguments{
\item{x}{S4 object of class \code{Binalysis} or \code{MetaboProfile}}
}
\value{
S4 object of class
}
\description{
Detect pre-treatment parameters for \code{Binalysis} or \code{MetaboProfile} class objects.
}
\examples{
## Retreive example file paths and sample information 
file_paths <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes') \%>\% 
   .[61:63]

sample_information <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes') \%>\% 
    dplyr::filter(name == 'QC01' | name == 'QC02' | name == 'QC03')

## Detect spectral binning parameters
bp <- binneR::detectParameters(file_paths)

## Perform spectral binning
bd <- binneR::binneRlyse(file_paths,sample_information,bp)

## Detect pre-treatment parameters
pp <- detectPretreatmentParameters(bd) 

pp
}
