% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportData}
\alias{exportData}
\alias{exportData,Binalysis-method}
\alias{exportData,MetaboProfile-method}
\alias{exportData,Analysis-method}
\alias{exportData,Assignment-method}
\alias{exportSampleInfo}
\alias{exportSampleInfo,Binalysis-method}
\alias{exportSampleInfo,MetaboProfile-method}
\alias{exportSampleInfo,Analysis-method}
\alias{exportAccurateData}
\alias{exportAccurateData,Binalysis-method}
\alias{exportPeakInfo}
\alias{exportPeakInfo,MetaboProfile-method}
\alias{exportModellingMetrics}
\alias{exportModellingMetrics,Analysis-method}
\alias{exportModellingImportance}
\alias{exportModellingImportance,Analysis-method}
\alias{exportModelling}
\alias{exportModelling,Analysis-method}
\alias{exportCorrelations}
\alias{exportCorrelations,Analysis-method}
\alias{exportAssignments}
\alias{exportAssignments,Assignment-method}
\alias{exportSummarisedAssignments}
\alias{exportSummarisedAssignments,Assignment-method}
\alias{exportConstruction}
\alias{exportConstruction,Construction-method}
\alias{exportSummarisedConstruction}
\alias{exportSummarisedConstruction,Construction-method}
\alias{export}
\alias{export,Binalysis-method}
\alias{export,MetaboProfile-method}
\alias{export,Analysis-method}
\alias{export,Assignment-method}
\alias{export,Construction-method}
\title{Export results}
\usage{
exportData(x, outPath = ".", ...)

\S4method{exportData}{Binalysis}(x, outPath = ".")

\S4method{exportData}{MetaboProfile}(x, outPath = ".")

\S4method{exportData}{Analysis}(x, outPath = ".", type = "raw", idx = "name")

\S4method{exportData}{Assignment}(x, outPath = ".")

exportSampleInfo(x, outPath = ".", ...)

\S4method{exportSampleInfo}{Binalysis}(x, outPath = ".")

\S4method{exportSampleInfo}{MetaboProfile}(x, outPath = ".")

\S4method{exportSampleInfo}{Analysis}(x, outPath = ".", type = "raw")

exportAccurateData(x, outPath = ".")

\S4method{exportAccurateData}{Binalysis}(x, outPath = ".")

exportPeakInfo(x, outPath = ".")

\S4method{exportPeakInfo}{MetaboProfile}(x, outPath = ".")

exportModellingMetrics(x, outPath = ".")

\S4method{exportModellingMetrics}{Analysis}(x, outPath = ".")

exportModellingImportance(x, outPath = ".")

\S4method{exportModellingImportance}{Analysis}(x, outPath = ".")

exportModelling(x, outPath = ".")

\S4method{exportModelling}{Analysis}(x, outPath = ".")

exportCorrelations(x, outPath = ".")

\S4method{exportCorrelations}{Analysis}(x, outPath = ".")

exportAssignments(x, outPath = ".")

\S4method{exportAssignments}{Assignment}(x, outPath = ".")

exportSummarisedAssignments(x, outPath = ".")

\S4method{exportSummarisedAssignments}{Assignment}(x, outPath = ".")

exportConstruction(x, outPath = ".")

\S4method{exportConstruction}{Construction}(x, outPath = ".")

exportSummarisedConstruction(x, outPath = ".")

\S4method{exportSummarisedConstruction}{Construction}(x, outPath = ".")

export(x, outPath = ".", ...)

\S4method{export}{Binalysis}(x, outPath = ".")

\S4method{export}{MetaboProfile}(x, outPath = ".")

\S4method{export}{Analysis}(x, outPath = ".", type = "raw", idx = "name")

\S4method{export}{Assignment}(x, outPath = ".")

\S4method{export}{Construction}(x, outPath = ".")
}
\arguments{
\item{x}{S4 object of class \code{Binalysis}, \code{MetaboProfile}, \code{Analysis} or \code{Assignment},}

\item{outPath}{directory path to export to.}

\item{...}{arguments to pass to relevant method}

\item{type}{data type to extract. \code{raw} or \code{pre-treated}}

\item{idx}{sample information column name to use as sample IDs}
}
\value{
A character vector of exported file paths.
}
\description{
Export data tables from \code{Binalysis},\code{MetaboProfile}, \code{Analysis} and \code{Assignment} classes.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

## Export spectrally binned data
export(analysis,outPath = tempdir())

## Perform data pre-treatment and modelling
p <- analysisParameters(c('pre-treatment','modelling'))
metabolyseR::parameters(p,'pre-treatment') <- preTreatmentParameters(
   list(occupancyFilter = 'maximum',
        transform = 'TICnorm')
)
metabolyseR::parameters(p,'modelling') <- metabolyseR::modellingParameters('anova')
metabolyseR::changeParameter(p,'cls') <- 'day'
analysis <- metabolyseR::metabolyse(metaboData::abr1$neg[,1:200],
                      metaboData::abr1$fact,
                      p)
                      
## Export pre-treated data and modelling results
export(analysis,outPath = tempdir())

## Perform molecular formula assignment
future::plan(future::sequential)
p <- assignments::assignmentParameters('FIE-HRMS')
assignments <- assignments::assignMFs(assignments::feature_data,p)

## Export molecular formula assignment results
export(assignments,outPath = tempdir())

## Perform consensus structural classification
structural_classifications <- construction::construction(assignments)

## Export consensus structural classification results
export(structural_classifications,outPath = tempdir())
}
