% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/featureSummary.R
\name{featureSummary}
\alias{featureSummary}
\alias{featureSummary,Binalysis-method}
\title{Summarise features}
\usage{
featureSummary(x)

\S4method{featureSummary}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\description{
Summarise spectral binning features.
}
