% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{featureSummary}
\alias{featureSummary}
\alias{featureSummary,Binalysis-method}
\title{Summarise features}
\usage{
featureSummary(x)

\S4method{featureSummary}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class \code{Binalysis}}
}
\value{
A tibble containing feature summaries
}
\description{
Summarise spectrally binned features.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

featureSummary(analysis)
}
