% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRSD.R
\name{plotRSD,Binalysis-method}
\alias{plotRSD,Binalysis-method}
\alias{plotRSD,MetaboProfile-method}
\title{Plot relative standard deviation (RSD) distributions}
\usage{
\S4method{plotRSD}{Binalysis}(analysis, cls = "class")

\S4method{plotRSD}{MetaboProfile}(analysis, cls = "class")
}
\arguments{
\item{analysis}{S4 object of class \code{Binalysis} or \code{MetaboProfile}}

\item{cls}{info column to use for class labels}

\item{QCidx}{QC sample label}

\item{QCparameters}{alternative parameters for QC sample pre-treatment. See details}

\item{histBins}{number of bins to use for histogram plotting}
}
\description{
Plot RSD distributions of processed data in quality control samples.
}
\details{
If \code{QCparameters} is set as NULL, the default QC pre-treatment parameters are used as given by \code{analysisParameters('preTreat')}. Alternative pre-treatment routines can be used by specifying an \code{AnalysisParameters} object for \code{QCparameters}.
}
