% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRSD.R
\name{plotRSD,Binalysis-method}
\alias{plotRSD,Binalysis-method}
\alias{plotRSD,MetaboProfile-method}
\title{Plot relative standard deviation (RSD) distributions}
\usage{
\S4method{plotRSD}{Binalysis}(analysis, cls = "class")

\S4method{plotRSD}{MetaboProfile}(analysis, cls = "class")
}
\arguments{
\item{analysis}{S4 object of class \code{Binalysis} or \code{MetaboProfile}}

\item{cls}{info column to use for class labels.}
}
\value{
A list of plots of RSD distributions
}
\description{
Plot RSD distributions of processed data.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))

## Plot RSD distributions
plotRSD(analysis)
}
