% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/preTreatModes.R
\docType{methods}
\name{preTreatModes}
\alias{preTreatModes}
\alias{preTreatModes,Binalysis-method}
\alias{preTreatModes,MetaboProfile-method}
\title{preTreatModes}
\usage{
preTreatModes(processedData, parameters, verbose = T)

\S4method{preTreatModes}{Binalysis}(processedData, parameters,
  verbose = T)

\S4method{preTreatModes}{MetaboProfile}(processedData, parameters,
  verbose = T)
}
\description{
preTreatModes
}
