% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-treat.R
\name{preTreatModes}
\alias{preTreatModes}
\alias{preTreatModes,Binalysis-method}
\alias{preTreatModes,MetaboProfile-method}
\title{Pre-treatment of multiple ionisation modes}
\usage{
preTreatModes(processed_data, parameters, verbose = TRUE)

\S4method{preTreatModes}{Binalysis}(processed_data, parameters, verbose = TRUE)

\S4method{preTreatModes}{MetaboProfile}(processed_data, parameters, verbose = TRUE)
}
\arguments{
\item{processed_data}{an object of S4 class \code{Binalysis} or \code{MetaboProfile}}

\item{parameters}{an object of S4 class \code{AnalysisParameters} containing pre-treatment parameters}

\item{verbose}{print console output}
}
\value{
An object of S4 class \code{Analysis}.
}
\description{
Pre-treat both positive and negative ionisation modes for \code{Binalysis} and \code{MetaboProfile} classes.
}
\examples{
## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Detect spectral binning parameters
bp <- binneR::detectParameters(files)

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = bp)

## Dectect the pre-treatment parameters
pre_treatment_parameters <- detectPretreatmentParameters(analysis)
metabolyseR::changeParameter(pre_treatment_parameters,'parallel') <- 'no'

## Perform pre-treatment
pre_treated_data <- preTreatModes(analysis,
                                  pre_treatment_parameters)
}
