% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preTreatModes.R
\name{preTreatModes}
\alias{preTreatModes}
\alias{preTreatModes,Binalysis-method}
\alias{preTreatModes,MetaboProfile-method}
\title{Pre-treatment of multiple ionisation modes}
\usage{
preTreatModes(processed_data, parameters, verbose = TRUE)

\S4method{preTreatModes}{Binalysis}(processed_data, parameters, verbose = TRUE)

\S4method{preTreatModes}{MetaboProfile}(processed_data, parameters, verbose = TRUE)
}
\arguments{
\item{processed_data}{S4 object of class \code{Binalysis} or \code{MetaboProfile}}

\item{parameters}{object of class \code{AnalysisParameters} containing pre-treatment parameters}

\item{verbose}{console output}
}
\value{
S4 object of class \code{Analysis}
}
\description{
Pre-treat both positive and negative ionisation modes for \code{Binalysis} and \code{MetaboProfile} classes.
}
\examples{
library(metabolyseR)

## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Detect spectral binning parameters
bp <- binneR::detectParameters(files)

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = bp)

## Declare pre-treatment parameters
pre_treatment_parameters <- analysisParameters('pre-treatment')
parameters(pre_treatment_parameters,
                        'pre-treatment') <- preTreatmentParameters(
 list(
   occupancyFilter = 'maximum',
   impute = 'all',
   transform = 'TICnorm'
 )
)
changeParameter(pre_treatment_parameters,'parallel') <- 'no'

## Perform pre-treatment
pre_treated_data <- preTreatModes(analysis,
                                  pre_treatment_parameters)
}
