% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/preTreatModes.R
\name{preTreatModes}
\alias{preTreatModes}
\alias{preTreatModes,Binalysis-method}
\alias{preTreatModes,MetaboProfile-method}
\title{preTreatModes}
\usage{
preTreatModes(processedData, parameters, verbose = TRUE)

\S4method{preTreatModes}{Binalysis}(processedData, parameters, verbose = T)

\S4method{preTreatModes}{MetaboProfile}(processedData, parameters, verbose = T)
}
\arguments{
\item{processedData}{of class Binalysis or MetaboProfile}

\item{parameters}{object of class AnalysisParameters containing pre-treatment parameters}

\item{verbose}{console output}
}
\description{
Pre-treat both positive and negative ionisation modes
}
