% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/reduce.R
\docType{methods}
\name{reduce}
\alias{reduce}
\alias{reduce,Analysis-method}
\title{reduce}
\usage{
reduce(x, isotopes = T, adducts = T, unknowns = F)

\S4method{reduce}{Analysis}(x, isotopes = T, adducts = T,
  unknowns = F)
}
\arguments{
\item{x}{S4 class of class Analysis}

\item{isotopes}{TRUE/FALSE remove isotopic features.}

\item{adducts}{TRUE/FALSE remove multiple adduct features.}

\item{unknowns}{TRUE/FALSE remove unassigned features.}
}
\description{
reduce an analysis by removing isotopic, adduct or unknown features.
}
\details{
Isotope and adduct features are filtered based on the maximum intensity peak for each molecular formulas.
}
