% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\alias{reduce,Analysis-method}
\alias{reduce,AnalysisData-method}
\title{Reduce an analysis by removing isotopic, adduct or unknown features}
\usage{
reduce(x, isotopes = TRUE, adducts = TRUE, unknowns = FALSE)

\S4method{reduce}{Analysis}(x, isotopes = TRUE, adducts = TRUE, unknowns = FALSE)

\S4method{reduce}{AnalysisData}(x, isotopes = TRUE, adducts = TRUE, unknowns = FALSE)
}
\arguments{
\item{x}{S4 class of class \code{Analysis} or \code{AnalysisData}}

\item{isotopes}{TRUE/FALSE remove isotopic features.}

\item{adducts}{TRUE/FALSE remove multiple adduct features.}

\item{unknowns}{TRUE/FALSE remove unassigned features.}
}
\value{
S4 object of class \code{Analysis} or \code{AnalysisData} with features reduced
}
\description{
Reduce an analysis by removing isotopic, adduct or unknown features.
}
\details{
Isotope and adduct features are filtered based on the maximum intensity peak for each molecular formulas.
}
\examples{
library(MFassign)

## Assign molecular formulas
p <- assignmentParameters('FIE')

assignment <- assignMFs(peakData,p)

## Retrieve assigned data
assigned_data <- metabolyseR::analysisData(
 assignedData(assignment),
 tibble::tibble(sample = seq_len(nrow(peakData)))
 )

reduced_data <- reduce(assigned_data)

reduced_data
}
