% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{suitableParallelPlan}
\alias{suitableParallelPlan}
\title{Generate a suitable parallel processing plan}
\usage{
suitableParallelPlan()
}
\description{
Generate a suitable parallel processing future plan for your session. See ?future::plan for more details on future plans?
}
\examples{
\dontrun{
suitableParallelPlan()
}
}
